Ignore windows that are not put on the pager
Previously, windows such as the start menu and calendar were treated as
normal windows. As a result, opening the start menu on a desktop edge
would cause a weird desktop animation and sometimes crash kwin.

Now, any window that does not want to be part of the pager is ignored
when deciding if a new desktop should be created on either end.